Refactor solver schemes
