+++ source/simulator/core.cc
@@ -1722,97 +1722,6 @@ namespace aspect
     global_volume = GridTools::volume (triangulation, *mapping);
   }
 
-  /**
-   * Converts a function with a certain number of components into a Function@dim@
-   * with optionally having additional zero components.
-   **/
-  template <int dim>
-  class VectorFunctionFromVectorFunctionObject : public Function<dim>
-  {
-    public:
-      /**
-       * Converts a function with @p n_object_components components into a Function@dim@
-       * while optionally providing additional components that are set to zero.
-       *
-       * @param function_object The function that will form the components
-       *     of the resulting Function object.
-       * @param first_component The first component that should be
-       *     filled.
-       * @param n_object_components The number of components that should be
-       *     filled from the first.
-       * @param n_total_components The total number of vector components of the
-       *     resulting Function object.
-       **/
-
-      VectorFunctionFromVectorFunctionObject (const std_cxx1x::function<void (const Point<dim> &,Vector<double> &)> &function_object,
-                                              const unsigned int first_component,
-                                              const unsigned int n_object_components,
-                                              const unsigned int n_total_components)
-        :
-        Function<dim>(n_total_components),
-        function_object (function_object),
-        first_component (first_component),
-        n_object_components (n_object_components)
-      {
-        Assert ((n_object_components > 0
-                 &&
-                 first_component+n_object_components <= n_total_components),
-                ExcMessage ("Number of objects components needs to be less than number of total components"));
-      }
-
-      double
-      value (const Point<dim> &p,
-             const unsigned int component) const
-      {
-        Assert (component < this->n_components,
-                ExcIndexRange (component, 0, this->n_components));
-
-        if (component < first_component)
-          return 0;
-        else if (component >= first_component + n_object_components)
-          return 0;
-        else
-          {
-            Vector<double> temp(n_object_components);
-            function_object (p, temp);
-            return temp(component - first_component);
-          }
-      }
-
-      void
-      vector_value (const Point<dim>   &p,
-                    Vector<double>     &values) const
-      {
-        AssertDimension(values.size(), this->n_components);
-
-        // set everything to zero, and then the right components to their correct values
-        values = 0;
-        Vector<double> temp(n_object_components);
-        function_object (p, temp);
-        for (unsigned int i = 0; i < n_object_components; i++)
-          {
-            values(first_component + i) = temp(i);
-          }
-      }
-
-    private:
-      /**
-       * The function object which we call when this class's solution() function is called.
-       **/
-      const std_cxx1x::function<void (const Point<dim> &,Vector<double> &)> function_object;
-
-      /**
-       * The first vector component whose value is to be filled by the given
-       * function.
-       */
-      const unsigned int first_component;
-      /**
-       * The number of vector components whose values are to be filled by the given
-       * function.
-       */
-      const unsigned int n_object_components;
-
-  };
 
 
   template <int dim>
@@ -1838,696 +1747,53 @@ namespace aspect
         current_linearization_point = distr_solution;
       }
 
+    // The free surface scheme is currently not build to work inside a nonlinear solver.
+    // We do the free surface execution at the beginning of the timestep for a specific reason.
+    // The time step size is calculated AFTER the whole solve_timestep() function.  If we call
+    // free_surface_execute() after the Stokes solve, it will be before we know what the appropriate
+    // time step to take is, and we will timestep the boundary incorrectly.
+    if (parameters.free_surface_enabled)
+      free_surface->execute ();
+
+    // Compute the reactions of compositional fields and temperature in case of operator splitting.
+    if (parameters.use_operator_splitting)
+      compute_reactions ();
+
     switch (parameters.nonlinear_solver)
       {
         case NonlinearSolver::IMPES:
         {
-          // We do the free surface execution at the beginning of the timestep for a specific reason.
-          // The time step size is calculated AFTER the whole solve_timestep() function.  If we call
-          // free_surface_execute() after the Stokes solve, it will be before we know what the appropriate
-          // time step to take is, and we will timestep the boundary incorrectly.
-          if (parameters.free_surface_enabled)
-            free_surface->execute ();
-
-          // then we compute the reactions of compositional fields and temperature in case of operator splitting
-          if (parameters.use_operator_splitting)
-            compute_reactions ();
-
-          assemble_advection_system (AdvectionField::temperature());
-          solve_advection(AdvectionField::temperature());
-
-          current_linearization_point.block(introspection.block_indices.temperature)
-            = solution.block(introspection.block_indices.temperature);
-
-          for (unsigned int c=0; c < introspection.n_compositional_fields; ++c)
-            {
-              const AdvectionField adv_field (AdvectionField::composition(c));
-              const typename Parameters<dim>::AdvectionFieldMethod::Kind method = adv_field.advection_method(introspection);
-              switch (method)
-                {
-                  case Parameters<dim>::AdvectionFieldMethod::fem_field:
-                    assemble_advection_system (adv_field);
-                    solve_advection(adv_field);
-                    break;
-
-                  case Parameters<dim>::AdvectionFieldMethod::particles:
-                    interpolate_particle_properties(adv_field);
-                    break;
-
-                  default:
-                    AssertThrow(false,ExcNotImplemented());
-                }
-            }
-
-          for (unsigned int c=0; c<introspection.n_compositional_fields; ++c)
-            current_linearization_point.block(introspection.block_indices.compositional_fields[c])
-              = solution.block(introspection.block_indices.compositional_fields[c]);
-
-          // the Stokes matrix depends on the viscosity. if the viscosity
-          // depends on other solution variables, then after we need to
-          // update the Stokes matrix in every time step and so need to set
-          // the following flag. if we change the Stokes matrix we also
-          // need to update the Stokes preconditioner.
-          if (stokes_matrix_depends_on_solution() == true)
-            rebuild_stokes_matrix = rebuild_stokes_preconditioner = true;
-
-          assemble_stokes_system();
-          build_stokes_preconditioner();
-          solve_stokes();
-
-          if (parameters.run_postprocessors_on_nonlinear_iterations)
-            postprocess ();
-
+          solve_IMPES();
           break;
         }
 
         case NonlinearSolver::Stokes_only:
         {
-          double initial_stokes_residual = 0.0;
-
-          do
-            {
-              // the Stokes matrix depends on the viscosity. if the viscosity
-              // depends on other solution variables, then we need to
-              // update the Stokes matrix in every iteration and so need to set
-              // the rebuild_stokes_matrix flag. if we change the Stokes matrix we also
-              // need to update the Stokes preconditioner.
-              //
-              // there is a similar case where this nonlinear solver can be used, namely for
-              // compressible models. in that case, the matrix does not depend on
-              // the previous solution, but we still need to iterate since the right
-              // hand side depends on it. in those cases, the matrix does not change,
-              // but if we have to repeat computing the right hand side, we need to
-              // also rebuild the matrix if we end up with inhomogeneous velocity
-              // boundary conditions (i.e., if there are prescribed velocity boundary
-              // indicators)
-              if ((stokes_matrix_depends_on_solution() == true)
-                  ||
-                  (parameters.prescribed_velocity_boundary_indicators.size() > 0))
-                rebuild_stokes_matrix = true;
-              if (stokes_matrix_depends_on_solution() == true)
-                rebuild_stokes_preconditioner = true;
-
-              assemble_stokes_system();
-              build_stokes_preconditioner();
-
-              if (nonlinear_iteration==0)
-                initial_stokes_residual = compute_initial_stokes_residual();
-
-              const double stokes_residual = solve_stokes();
-
-              const double relative_tolerance = (initial_stokes_residual > 0) ? stokes_residual/initial_stokes_residual : 0.0;
-
-              pcout << "      Relative Stokes residual after nonlinear iteration " << nonlinear_iteration+1
-                    << ": " << relative_tolerance
-                    << std::endl;
-
-              if (parameters.run_postprocessors_on_nonlinear_iterations)
-                postprocess ();
-
-              if (relative_tolerance < parameters.nonlinear_tolerance)
-                break;
-
-              current_linearization_point = solution;
-
-              ++nonlinear_iteration;
-            }
-          while (! ((nonlinear_iteration >= parameters.max_nonlinear_iterations) // regular timestep
-                    ||
-                    ((pre_refinement_step < parameters.initial_adaptive_refinement) // pre-refinement
-                     &&
-                     (nonlinear_iteration >= parameters.max_nonlinear_iterations_in_prerefinement))));
+          solve_Stokes_only();
           break;
         }
 
-
         case NonlinearSolver::iterated_IMPES:
         {
-          double initial_temperature_residual = 0;
-          double initial_stokes_residual      = 0;
-          std::vector<double> initial_composition_residual (introspection.n_compositional_fields,0);
-
-          // compute the reactions of compositional fields and temperature in case of operator splitting
-          if (parameters.use_operator_splitting)
-            compute_reactions ();
-
-          do
-            {
-              assemble_advection_system(AdvectionField::temperature());
-
-              if (nonlinear_iteration == 0)
-                initial_temperature_residual = system_rhs.block(introspection.block_indices.temperature).l2_norm();
-
-              const double temperature_residual = solve_advection(AdvectionField::temperature());
-              const double relative_temperature_residual = (initial_temperature_residual > 0)
-                                                           ?
-                                                           temperature_residual/initial_temperature_residual
-                                                           :
-                                                           0.0;
-
-              current_linearization_point.block(introspection.block_indices.temperature)
-                = solution.block(introspection.block_indices.temperature);
-              rebuild_stokes_matrix = true;
-              std::vector<double> relative_composition_residual (introspection.n_compositional_fields,0);
-
-              for (unsigned int c=0; c<introspection.n_compositional_fields; ++c)
-                {
-                  const AdvectionField adv_field (AdvectionField::composition(c));
-                  typename Parameters<dim>::AdvectionFieldMethod::Kind method = adv_field.advection_method(introspection);
-                  switch (method)
-                    {
-                      case Parameters<dim>::AdvectionFieldMethod::fem_field:
-                      {
-                        assemble_advection_system (adv_field);
-
-                        if (nonlinear_iteration == 0)
-                          initial_composition_residual[c] = system_rhs.block(introspection.block_indices.compositional_fields[c]).l2_norm();
-
-                        const double composition_residual
-                          = solve_advection(adv_field);
-
-                        relative_composition_residual[c] = (initial_composition_residual[c] > 0)
-                                                           ?
-                                                           composition_residual/initial_composition_residual[c]
-                                                           :
-                                                           0.0;
-                      }
-                      break;
-
-                      case Parameters<dim>::AdvectionFieldMethod::particles:
-                        interpolate_particle_properties(adv_field);
-                        break;
-
-                      default:
-                        AssertThrow(false,ExcNotImplemented());
-                    }
-                }
-
-              // for consistency we update the current linearization point only after we have solved
-              // all fields, so that we use the same point in time for every field when solving
-              for (unsigned int c=0; c<introspection.n_compositional_fields; ++c)
-                current_linearization_point.block(introspection.block_indices.compositional_fields[c])
-                  = solution.block(introspection.block_indices.compositional_fields[c]);
-
-              // the Stokes matrix depends on the viscosity. if the viscosity
-              // depends on other solution variables, then after we need to
-              // update the Stokes matrix in every time step and so need to set
-              // the following flag. if we change the Stokes matrix we also
-              // need to update the Stokes preconditioner.
-              if (stokes_matrix_depends_on_solution() == true)
-                rebuild_stokes_matrix = rebuild_stokes_preconditioner = true;
-
-              assemble_stokes_system();
-              build_stokes_preconditioner();
-
-              if (nonlinear_iteration == 0)
-                initial_stokes_residual = compute_initial_stokes_residual();
-
-              const double stokes_residual = solve_stokes();
-              const double relative_stokes_residual = (initial_stokes_residual > 0) ? stokes_residual/initial_stokes_residual : 0.0;
-
-              current_linearization_point = solution;
-
-              // write the residual output in the same order as the output when
-              // solving the equations, output only relative residuals
-              pcout << "      Relative nonlinear residuals: " << relative_temperature_residual;
-              for (unsigned int c=0; c<introspection.n_compositional_fields; ++c)
-                pcout << ", " << relative_composition_residual[c];
-              pcout << ", " << relative_stokes_residual;
-              pcout << std::endl;
-
-              double max = 0.0;
-              for (unsigned int c=0; c<introspection.n_compositional_fields; ++c)
-                {
-                  // in models with melt migration the melt advection equation includes the divergence of the velocity
-                  // and can not be expected to converge to a smaller value than the residual of the Stokes equation.
-                  // thus, we set a threshold for the initial composition residual.
-                  // this only plays a role if the right-hand side of the advection equation is very small.
-                  const double threshold = (parameters.include_melt_transport && c == introspection.compositional_index_for_name("porosity")
-                                            ?
-                                            parameters.linear_stokes_solver_tolerance * time_step
-                                            :
-                                            0.0);
-                  if (initial_composition_residual[c]>threshold)
-                    max = std::max(relative_composition_residual[c],max);
-                }
-
-              max = std::max(relative_stokes_residual, max);
-              max = std::max(relative_temperature_residual, max);
-              pcout << "      Total relative residual after nonlinear iteration " << nonlinear_iteration+1 << ": " << max << std::endl;
-              pcout << std::endl
-                    << std::endl;
-
-              if (parameters.run_postprocessors_on_nonlinear_iterations)
-                postprocess ();
-
-              if (max < parameters.nonlinear_tolerance)
-                break;
-
-              ++nonlinear_iteration;
-
-//TODO: terminate here if the number of iterations is too large and we see no convergence
-            }
-          while (! ((nonlinear_iteration >= parameters.max_nonlinear_iterations) // regular timestep
-                    ||
-                    ((pre_refinement_step < parameters.initial_adaptive_refinement) // pre-refinement
-                     &&
-                     (nonlinear_iteration >= parameters.max_nonlinear_iterations_in_prerefinement))));
-
+          solve_iterated_IMPES();
           break;
         }
 
         case NonlinearSolver::iterated_Stokes:
         {
-          if (parameters.free_surface_enabled)
-            free_surface->execute ();
-
-          // then we compute the reactions of compositional fields and temperature in case of operator splitting
-          if (parameters.use_operator_splitting)
-            compute_reactions ();
-
-          // solve the temperature and composition systems once...
-          assemble_advection_system (AdvectionField::temperature());
-          solve_advection(AdvectionField::temperature());
-          current_linearization_point.block(introspection.block_indices.temperature)
-            = solution.block(introspection.block_indices.temperature);
-
-          for (unsigned int c=0; c<introspection.n_compositional_fields; ++c)
-            {
-              const AdvectionField adv_field (AdvectionField::composition(c));
-              typename Parameters<dim>::AdvectionFieldMethod::Kind method = adv_field.advection_method(introspection);
-              switch (method)
-                {
-                  case Parameters<dim>::AdvectionFieldMethod::fem_field:
-                    assemble_advection_system (adv_field);
-                    solve_advection(adv_field);
-                    break;
-
-                  case Parameters<dim>::AdvectionFieldMethod::particles:
-                    interpolate_particle_properties(adv_field);
-                    break;
-
-                  default:
-                    AssertThrow(false,ExcNotImplemented());
-                }
-            }
-
-          for (unsigned int c=0; c<introspection.n_compositional_fields; ++c)
-            current_linearization_point.block(introspection.block_indices.compositional_fields[c])
-              = solution.block(introspection.block_indices.compositional_fields[c]);
-
-          // residual vector (only for the velocity)
-          LinearAlgebra::Vector residual (introspection.index_sets.system_partitioning[0], mpi_communicator);
-          LinearAlgebra::Vector tmp (introspection.index_sets.system_partitioning[0], mpi_communicator);
-
-          // ...and then iterate the solution of the Stokes system
-          double initial_stokes_residual = 0;
-          for (nonlinear_iteration=0; (! ((nonlinear_iteration >= parameters.max_nonlinear_iterations) // regular timestep
-                                          ||
-                                          ((pre_refinement_step < parameters.initial_adaptive_refinement) // pre-refinement
-                                           &&
-                                           (nonlinear_iteration >= parameters.max_nonlinear_iterations_in_prerefinement)))); ++nonlinear_iteration)
-            {
-              // rebuild the matrix if it actually depends on the solution
-              // of the previous iteration.
-              if ((stokes_matrix_depends_on_solution() == true)
-                  ||
-                  (parameters.prescribed_velocity_boundary_indicators.size() > 0))
-                rebuild_stokes_matrix = rebuild_stokes_preconditioner = true;
-
-              assemble_stokes_system();
-              build_stokes_preconditioner();
-
-              if (nonlinear_iteration==0)
-                initial_stokes_residual = compute_initial_stokes_residual();
-
-              const double stokes_residual = solve_stokes();
-              const double relative_stokes_residual = (initial_stokes_residual > 0) ? stokes_residual/initial_stokes_residual : 0.0;
-
-              pcout << "      Relative Stokes residual after nonlinear iteration " << nonlinear_iteration+1
-                    << ": " << relative_stokes_residual
-                    << std::endl;
-
-              if (parameters.run_postprocessors_on_nonlinear_iterations)
-                postprocess ();
-
-              if (stokes_residual/initial_stokes_residual < parameters.nonlinear_tolerance)
-                {
-                  break; // convergence reached, exit nonlinear iterations.
-                }
-
-              current_linearization_point.block(introspection.block_indices.velocities)
-                = solution.block(introspection.block_indices.velocities);
-              if (introspection.block_indices.velocities != introspection.block_indices.pressure)
-                current_linearization_point.block(introspection.block_indices.pressure)
-                  = solution.block(introspection.block_indices.pressure);
-
-              pcout << std::endl;
-            }
-
+          solve_iterated_Stokes();
           break;
         }
 
         case NonlinearSolver::Newton_Stokes:
         {
-          if (parameters.free_surface_enabled)
-            free_surface->execute ();
-
-          std::vector<double> initial_composition_residual (parameters.n_compositional_fields,0);
-
-          double initial_residual = 1;
-
-          double velocity_residual = 0;
-          double pressure_residual = 0;
-          double residual = 1;
-          double residual_old = 1;
-
-          double switch_initial_residual = 1;
-          double newton_residual_for_derivative_scaling_factor = 1;
-
-          bool use_picard = true;
-
-          const unsigned int max_iterations = (pre_refinement_step < parameters.initial_adaptive_refinement)
-                                              ?
-                                              std::min (parameters.max_nonlinear_iterations,parameters.max_nonlinear_iterations_in_prerefinement)
-                                              :
-                                              parameters.max_nonlinear_iterations;
-
-          double stokes_residual = 0;
-          for (nonlinear_iteration = 0; nonlinear_iteration < max_iterations; ++nonlinear_iteration)
-            {
-              assemble_newton_stokes_system = true;
-
-              solution.block(introspection.block_indices.pressure) = 0;
-              solution.block(introspection.block_indices.velocities) = 0;
-
-              assemble_advection_system(AdvectionField::temperature());
-              build_advection_preconditioner(AdvectionField::temperature(),
-                                             T_preconditioner);
-
-              solve_advection(AdvectionField::temperature());
-
-              current_linearization_point.block(introspection.block_indices.temperature)
-                = solution.block(introspection.block_indices.temperature);
-
-              std::vector<double> composition_residual (parameters.n_compositional_fields,0.0);
-
-              for (unsigned int c=0; c<parameters.n_compositional_fields; ++c)
-                {
-                  assemble_advection_system (AdvectionField::composition(c));
-                  build_advection_preconditioner(AdvectionField::composition(c),
-                                                 C_preconditioner);
-
-                  if (nonlinear_iteration == 0)
-                    initial_composition_residual[c] = system_rhs.block(introspection.block_indices.compositional_fields[c]).l2_norm();
-
-                  composition_residual[c]
-                    = solve_advection(AdvectionField::composition(c));
-                }
-
-              // for consistency we update the current linearization point only after we have solved
-              // all fields, so that we use the same point in time for every field when solving
-              for (unsigned int c=0; c<parameters.n_compositional_fields; ++c)
-                current_linearization_point.block(introspection.block_indices.compositional_fields[c])
-                  = solution.block(introspection.block_indices.compositional_fields[c]);
-
-              if (use_picard == true && (residual/initial_residual <= parameters.nonlinear_switch_tolerance ||
-                                         nonlinear_iteration >= parameters.max_pre_newton_nonlinear_iterations))
-                {
-                  use_picard = false;
-                  pcout << "   Switching from defect correction form of Picard to the Newton solver scheme." << std::endl;
-
-                  /**
-                   * This method allows to slowly introduce the derivatives based
-                   * on the improvement of the residual. If we do not use it, we
-                   * just set it so the newton_derivative_scaling_factor goes from
-                   * zero to one when switching on the Newton solver.
-                   */
-                  if (!parameters.use_newton_residual_scaling_method)
-                    newton_residual_for_derivative_scaling_factor = 0;
-                }
-
-              newton_handler->set_newton_derivative_scaling_factor(std::max(0.0,
-                                                                            (1.0-(newton_residual_for_derivative_scaling_factor/switch_initial_residual))));
-
-
-              /**
-               * copied from solver.cc
-               */
-
-              // Many parts of the solver depend on the block layout (velocity = 0,
-              // pressure = 1). For example the linearized_stokes_initial_guess vector or the StokesBlock matrix
-              // wrapper. Let us make sure that this holds (and shorten their names):
-              const unsigned int block_vel = introspection.block_indices.velocities;
-              const unsigned int block_p = (parameters.include_melt_transport) ?
-                                           introspection.variable("fluid pressure").block_index
-                                           : introspection.block_indices.pressure;
-              Assert(block_vel == 0, ExcNotImplemented());
-              Assert(block_p == 1, ExcNotImplemented());
-              Assert(!parameters.include_melt_transport
-                     || introspection.variable("compaction pressure").block_index == 1, ExcNotImplemented());
-
-              // create a completely distributed vector that will be used for
-              // the scaled and denormalized solution and later used as a
-              // starting guess for the linear solver
-              LinearAlgebra::BlockVector linearized_stokes_initial_guess (introspection.index_sets.stokes_partitioning, mpi_communicator);
-
-              linearized_stokes_initial_guess.block (block_vel) = current_linearization_point.block (block_vel);
-              linearized_stokes_initial_guess.block (block_p) = current_linearization_point.block (block_p);
-
-              if (nonlinear_iteration == 0)
-                {
-                  initial_residual = compute_initial_newton_residual(linearized_stokes_initial_guess);
-                  switch_initial_residual = initial_residual;
-                  residual = initial_residual;
-                }
-
-              assemble_newton_stokes_system = assemble_newton_stokes_matrix = true;
-
-              denormalize_pressure (last_pressure_normalization_adjustment,
-                                    linearized_stokes_initial_guess,
-                                    current_linearization_point);
-
-              if (nonlinear_iteration <= 1)
-                compute_current_constraints ();
-
-              // the Stokes matrix depends on the viscosity. if the viscosity
-              // depends on other solution variables, then after we need to
-              // update the Stokes matrix in every time step and so need to set
-              // the following flag. if we change the Stokes matrix we also
-              // need to update the Stokes preconditioner.
-              rebuild_stokes_matrix = rebuild_stokes_preconditioner = assemble_newton_stokes_matrix = true;
-
-              assemble_stokes_system();
-
-              /**
-               * Eisenstat Walker method for determining the tolerance
-               */
-              if (nonlinear_iteration > 1)
-                {
-                  residual_old = residual;
-                  velocity_residual = system_rhs.block(introspection.block_indices.velocities).l2_norm();
-                  pressure_residual = system_rhs.block(introspection.block_indices.pressure).l2_norm();
-                  residual = std::sqrt(velocity_residual * velocity_residual + pressure_residual * pressure_residual);
-
-                  if (!use_picard)
-                    {
-                      const bool EisenstatWalkerChoiceOne = true;
-                      parameters.linear_stokes_solver_tolerance = compute_Eisenstat_Walker_linear_tolerance(EisenstatWalkerChoiceOne,
-                                                                  parameters.maximum_linear_stokes_solver_tolerance,
-                                                                  parameters.linear_stokes_solver_tolerance,
-                                                                  stokes_residual,
-                                                                  residual,
-                                                                  residual_old);
-
-                      pcout << "   The linear solver tolerance is set to " << parameters.linear_stokes_solver_tolerance << std::endl;
-                    }
-                }
-
-              build_stokes_preconditioner();
-
-              stokes_residual = solve_stokes();
-
-              velocity_residual = system_rhs.block(introspection.block_indices.velocities).l2_norm();
-              pressure_residual = system_rhs.block(introspection.block_indices.pressure).l2_norm();
-              residual = std::sqrt(velocity_residual * velocity_residual + pressure_residual * pressure_residual);
-
-              /**
-               * We may need to do a line search if the solution update doesn't decrease the norm of the rhs enough.
-               * This is done by adding the solution update to the current linearization point and then assembling
-               * the Newton right hand side. If the Newton residual has decreased enough by using the this update,
-               * then we continue, otherwise we reset the current linearization point with the help of the backup and
-               * add each iteration an increasingly smaller solution update until the decreasing residual condition
-               * is met, or the line search iteration limit is reached.
-               */
-              LinearAlgebra::BlockVector backup_linearization_point = current_linearization_point;
-
-              double test_residual = 0;
-              double test_velocity_residual = 0;
-              double test_pressure_residual = 0;
-              double lambda = 1;
-              double alpha = 1e-4;
-              unsigned int line_search_iteration = 0;
-
-              /**
-               * Do the loop for the line search. Even when we
-               * don't do a line search we go into this loop
-               */
-              do
-                {
-                  current_linearization_point = backup_linearization_point;
-
-                  LinearAlgebra::BlockVector search_direction = solution;
-
-                  search_direction *= lambda;
-
-
-                  current_linearization_point.block(introspection.block_indices.pressure) += search_direction.block(introspection.block_indices.pressure);
-                  current_linearization_point.block(introspection.block_indices.velocities) += search_direction.block(introspection.block_indices.velocities);
-
-                  assemble_newton_stokes_matrix = rebuild_stokes_preconditioner = false;
-                  rebuild_stokes_matrix = true;
-
-                  assemble_stokes_system();
-
-                  test_velocity_residual = system_rhs.block(introspection.block_indices.velocities).l2_norm();
-                  test_pressure_residual = system_rhs.block(introspection.block_indices.pressure).l2_norm();
-                  test_residual = std::sqrt(test_velocity_residual * test_velocity_residual
-                                            + test_pressure_residual * test_pressure_residual);
-
-                  if (test_residual < (1.0 - alpha * lambda) * residual
-                      ||
-                      line_search_iteration >= parameters.max_newton_line_search_iterations
-                      ||
-                      use_picard)
-                    {
-                      pcout << "      Total relative residual after nonlinear iteration " << nonlinear_iteration+1
-                            << ": " << residual/initial_residual << ", norm of the rhs: " << test_residual
-                            << ", newton_derivative_scaling_factor: " << newton_handler->get_newton_derivative_scaling_factor() << std::endl;
-                      break;
-                    }
-                  else
-                    {
-
-                      pcout << "   Line search iteration " << line_search_iteration << ", with norm of the rhs "
-                            << test_residual << " and going to " << (1.0 - alpha * lambda) * residual
-                            << ", relative residual: " << residual/initial_residual << std::endl;
-
-                      /**
-                       * The line search step was not sufficient to decrease the residual
-                       * enough, so we take a smaller step to see if it improves the residual.
-                       */
-                      lambda *= (2.0/3.0);// TODO: make a parameter out of this.
-                    }
-
-                  line_search_iteration++;
-                  Assert(line_search_iteration <= parameters.max_newton_line_search_iterations,
-                         ExcMessage ("This tests the while condition. This condition should "
-                                     "actually never be false, because the break statement "
-                                     "above should have caught it."));
-                }
-              while (line_search_iteration <= parameters.max_newton_line_search_iterations);
-              // The while condition should actually never be false, because the break statement above should have caught it.
-
-
-              if (use_picard == true)
-                {
-                  switch_initial_residual = residual;
-                  newton_residual_for_derivative_scaling_factor = residual;
-                }
-              else
-                {
-                  /**
-                   * This method allows to slowly introduce the derivatives based
-                   * on the improvement of the residual. This method was suggested
-                   * by Raid Hassani.
-                   */
-                  if (parameters.use_newton_residual_scaling_method)
-                    newton_residual_for_derivative_scaling_factor = test_residual;
-                  else
-                    newton_residual_for_derivative_scaling_factor = 0;
-                }
-
-              residual_old = residual;
-
-              pcout << std::endl;
-              last_pressure_normalization_adjustment = normalize_pressure(current_linearization_point);
-
-              if (residual/initial_residual < parameters.nonlinear_tolerance)
-                break;
-            }
-
-          // When we are finished iterating, we need to set the final solution to the current linearization point,
-          // because the solution vector is used in the postprocess.
-          solution = current_linearization_point;
+          solve_Newton_Stokes();
           break;
         }
 
         case NonlinearSolver::Advection_only:
         {
-          // Identical to IMPES except does not solve Stokes equation
-          if (parameters.free_surface_enabled)
-            free_surface->execute ();
-
-          // then we compute the reactions of compositional fields and temperature in case of operator splitting
-          if (parameters.use_operator_splitting)
-            compute_reactions ();
-
-          LinearAlgebra::BlockVector distributed_stokes_solution (introspection.index_sets.system_partitioning, mpi_communicator);
-
-          VectorFunctionFromVectorFunctionObject<dim> func(std_cxx1x::bind (&PrescribedStokesSolution::Interface<dim>::stokes_solution,
-                                                                            std_cxx1x::cref(*prescribed_stokes_solution),
-                                                                            std_cxx1x::_1,
-                                                                            std_cxx1x::_2),
-                                                           0,
-                                                           dim+1, // velocity and pressure
-                                                           introspection.n_components);
-
-          VectorTools::interpolate (*mapping, dof_handler, func, distributed_stokes_solution);
-
-          const unsigned int block_vel = introspection.block_indices.velocities;
-          const unsigned int block_p = introspection.block_indices.pressure;
-
-          // distribute hanging node and
-          // other constraints
-          current_constraints.distribute (distributed_stokes_solution);
-
-          solution.block(block_vel) = distributed_stokes_solution.block(block_vel);
-          solution.block(block_p) = distributed_stokes_solution.block(block_p);
-
-          assemble_advection_system (AdvectionField::temperature());
-          solve_advection(AdvectionField::temperature());
-
-          current_linearization_point.block(introspection.block_indices.temperature)
-            = solution.block(introspection.block_indices.temperature);
-
-          for (unsigned int c=0; c<introspection.n_compositional_fields; ++c)
-            {
-              const AdvectionField adv_field (AdvectionField::composition(c));
-              typename Parameters<dim>::AdvectionFieldMethod::Kind method = adv_field.advection_method(introspection);
-              switch (method)
-                {
-                  case Parameters<dim>::AdvectionFieldMethod::fem_field:
-                    assemble_advection_system (adv_field);
-                    solve_advection(adv_field);
-                    break;
-
-                  case Parameters<dim>::AdvectionFieldMethod::particles:
-                    interpolate_particle_properties(adv_field);
-                    break;
-
-                  default:
-                    AssertThrow(false,ExcNotImplemented());
-                }
-            }
-
-          for (unsigned int c=0; c<introspection.n_compositional_fields; ++c)
-            current_linearization_point.block(introspection.block_indices.compositional_fields[c])
-              = solution.block(introspection.block_indices.compositional_fields[c]);
-
-          if (parameters.run_postprocessors_on_nonlinear_iterations)
-            postprocess ();
-
+          solve_advection_only();
           break;
         }
